requirements.txt
streamlit==1.28.0
numpy==1.24.3
pillow==10.0.0
tensorflow==2.15.0
requests==2.31.0
matplotlib==3.7.2